+++ src/collective/iconifiedcategory/tests/test_behaviors.py
@@ -185,3 +185,55 @@ class TestIconifiedCategorization(BaseTestCase, unittest.TestCase):
         self.assertEqual(category2_brain.UID, obj.UID())
         obj_brain = catalog(UID=obj.UID())[0]
         self.assertEqual(obj_brain.content_category_uid, category2.UID())
+
+    def test_content_category_changed_default_values(self):
+        """While content_category is changed on an element, the default values for fields
+           to_print/confidential/to_sign/signed are reapplied with new content_category
+           default values if it was still the default vlue of the original content_category."""
+        category_group = self.portal.config['group-1']
+        category_group.to_be_printed_activated = True
+        category_group.confidentiality_activated = True
+        category_group.signed_activated = True
+        category11 = self.portal.config['group-1']['category-1-1']
+        category11_id = calculate_category_id(category11)
+        obj = api.content.create(
+            id='file2',
+            type='File',
+            file=self.file,
+            container=self.portal,
+            content_category=category11_id)
+        self.assertFalse(obj.confidential)
+        self.assertFalse(obj.to_sign)
+        self.assertFalse(obj.signed)
+        self.assertIsNone(obj.to_print)
+        # now enable everything on category-1-2 and use it
+        category12 = self.portal.config['group-1']['category-1-2']
+        category12.to_print = False
+        category12.confidential = True
+        category12.to_sign = True
+        category12.signed = True
+        category12_id = calculate_category_id(category12)
+        adapted_obj = IconifiedCategorization(obj)
+        setattr(adapted_obj, 'content_category', category12_id)
+        notify(ObjectModifiedEvent(obj))
+
+        # changed to True on obj an in parent's categorized_elements
+        parent_cat_elements = obj.aq_parent.categorized_elements[obj.UID()]
+        self.assertTrue(obj.confidential)
+        self.assertTrue(parent_cat_elements['confidential'])
+        self.assertTrue(obj.to_sign)
+        self.assertTrue(parent_cat_elements['to_sign'])
+        self.assertTrue(obj.signed)
+        self.assertTrue(parent_cat_elements['signed'])
+        # to_print could not be set to False because not printable
+        self.assertIsNone(obj.to_print)
+        self.assertIsNone(parent_cat_elements['to_print'])
+
+        # enable conversion and back to category11
+        category11.to_print = True
+        gsettings = GlobalSettings(self.portal)
+        gsettings.auto_layout_file_types = CONVERTABLE_TYPES.keys()
+        setattr(adapted_obj, 'content_category', category11_id)
+        notify(ObjectModifiedEvent(obj))
+        self.assertTrue(obj.to_print)
+        self.assertTrue(parent_cat_elements['to_print'])